+++ NEWS.md
@@ -27,6 +27,12 @@ New features:
   - `options("marginaleffects_posterior_center" = median)`
   - `options("marginaleffects_posterior_center" = mean)`
 
+Renamed functions (backward-compatibility is maintained):
+
+* `marginaleffects()` -> `slopes()` 
+* `posteriordraws()` -> `posterior_draws()` 
+* `marginalmeans()` -> `marginal_means()` 
+
 Bug fixes:
 
 * Incorrect results: In 0.8.1, `plot_c*()` the `threenum` and `minmax` labels did not correspond to the correct numeric values.
@@ -42,7 +48,7 @@ Other:
 
 * `deltamethod()` can run hypothesis tests on objects produced by the `comparisons()`, `marginaleffects()`, `predictions()`, and `marginalmeans()` functions. This feature relies on `match.call()`, which means it may not always work when used programmatically, inside functions and nested environments. It is generally safer and more efficient to use the `hypothesis` argument.
 * `plot_cme()` and `plot_cco()` accept lists with user-specified values for the regressors, and can display nice labels for shortcut string-functions like "threenum" or "quartile".
-* `posteriordraws`: new `shape` argument to return MCMC draws in various formats, including the new `rvar` structure from the `posterior` package.
+* `posterior_draws`: new `shape` argument to return MCMC draws in various formats, including the new `rvar` structure from the `posterior` package.
 * `transform_avg` function gets printed in `summary()` output.
 * `transform_post` and `transform_avg` support string shortcuts: "exp" and "ln"
 * Added support for `mlm` models from `lm()`. Thanks to Noah Greifer.
@@ -327,7 +333,7 @@ Misc:
 * `datagrid` function supersedes `typical` and `counterfactual` with the `grid.type`
   argument. The `typical` and `counterfactual` functions will remain available
   and exported, but their use is not encouraged.
-* `posteriordraws` function can be applied to a `predictions` or a
+* `posterior_draws` function can be applied to a `predictions` or a
   `marginaleffects` object to extract draws from the posterior distribution.
 * `marginalmeans` standard errors are now computed using the delta method.
 * `predictions` standard errors are now computed using the delta method when they are not available from `insight::get_predicted`.